requirements.txt
#TODO: add versions
lkml
database_mappers
errors
google-cloud-bigquery
argparse
logging
os
subprocess
sys
pathlib
